requirements.txt
# Core dependencies for AI Video Essay Pipeline
# Pin versions for reproducibility; adjust as needed for experimentation.
# Updated ADK dependency: use official PyPI package for stable release
google-adk
python-dotenv>=1.0
PyYAML>=6.0
pytest>=8.0
litellm
